+++ src/index.js
@@ -83,6 +83,12 @@ export { Labels } from "./components/Labels/Labels";
 export { Breadcrumb } from "./components/Breadcrumb/Breadcrumb";
 
 /**
+ * FormError
+ */
+
+export { FormError } from "./components/FormError/FormError";
+
+/**
  * Form Label
  */
 export { FormLabel } from "./components/FormLabel/FormLabel";